backend/requirements.txt
fastapi
uvicorn[standard]
python-docx
python-dotenv
google-generativeai
requests
redis
openpyxl
lxml
biopython
